+++ .github/workflows/chunk-ingest.yaml
@@ -13,7 +13,7 @@ on:
       - 'java/ingest/ingest-core/**'
       - 'java/ingest/ingest-tracker/**'
       - 'java/ingest/ingest-trackerV2/**'
-      - 'java/ingest/ingest-starter/**'
+      - 'java/ingest/ingest-taskrunner/**'
       - 'java/ingest/ingest-batcher-core/**'
       - 'java/ingest/ingest-batcher-store/**'
       - 'java/ingest/ingest-batcher-storev2/**'
