+++ java/ingest/ingest-runner/src/test/java/sleeper/ingest/runner/testutils/InMemoryPartitionFileWriter.java
@@ -29,7 +29,7 @@ import sleeper.core.statestore.FileReference;
 import sleeper.core.table.TableFilePaths;
 import sleeper.ingest.runner.impl.partitionfilewriter.PartitionFileWriter;
 import sleeper.ingest.runner.impl.partitionfilewriter.PartitionFileWriterFactory;
-import sleeper.sketches.Sketches;
+import sleeper.sketchesv2.Sketches;
 
 import java.util.ArrayList;
 import java.util.List;
