+++ java/ingest/ingest-runner/src/test/java/sleeper/ingest/runner/testutils/InMemorySketchesStore.java
@@ -15,7 +15,7 @@
  */
 package sleeper.ingest.runner.testutils;
 
-import sleeper.sketches.Sketches;
+import sleeper.sketchesv2.Sketches;
 
 import java.util.HashMap;
 import java.util.Map;
