+++ java/ingest/ingest-taskrunner/pom.xml
@@ -23,7 +23,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>ingest-starter</artifactId>
+    <artifactId>ingest-taskrunner</artifactId>
 
     <dependencies>
         <dependency>
