+++ java/ingest/ingest-taskrunner/src/main/java/sleeper/ingest/taskrunner/RunIngestTasksLambda.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.ingest.starter;
+package sleeper.ingest.taskrunner;
 
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.events.ScheduledEvent;
